backend/service_header/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
openpyxl==3.1.2
mysql-connector-python==8.0.27